+++ lib/__init__.py
@@ -2,8 +2,8 @@
 Library package for ACM switchover automation.
 """
 
-__version__ = "1.7.4"
-__version_date__ = "2026-04-24"
+__version__ = "1.7.5"
+__version_date__ = "2026-04-27"
 
 from .exceptions import (
     ConfigurationError,